apps/api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["liana <liana.soima@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.109.2"
uvicorn = "^0.27.1"
email-validator = "^2.1.0"
pydantic-settings = "^2.2.1"
python-dotenv = "^1.0.1"
supabase-py-async = "^2.5.5"
asyncpg = "^0.29.0"
gotrue = "2.8.1"

[tool.poetry.dev-dependencies]
isort = "5.10.1"
black = "22.6.0"
pytest = "8.0.1"
pylint-pydantic = "0.3.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
